+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 # [?.?.?] Unreleased
 
 ### Added
+- `say_info`, `say_verbose`, `say_debug`, `say_warn`, `say_crit`, `say_error`,
+  `say_fatal`, `say_sys_error` macros for basic logging into the tarantool's
+  default logger. Use these when you need to log messages, but don't have any
+  facilities set up for this, e.g. internally in tarantool-module code.
+- `log::SayLevel::current` function for getting current tarantool log level.
 
 ### Changed